requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclass-factory==2.16
Flask==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pika==1.3.2
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.4
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
